+++ FROST-Server.Core/src/main/java/de/fraunhofer/iosb/ilt/frostserver/settings/annotation/SensitiveValue.java
@@ -15,23 +15,19 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package de.fraunhofer.iosb.ilt.frostserver.settings;
+package de.fraunhofer.iosb.ilt.frostclient.settings.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
 
 /**
- * Thrown when a property is requested in a type that the value of the property
- * can not be converted to.
+ * An annotation specifying the field holds a security-sensitive value that must
+ * not be written to log files, like a password.
  */
-public class PropertyTypeException extends PropertyException {
-
-    public PropertyTypeException(String key, Class<?> typeExpected, Throwable cause) {
-        super(key + " has not expected type" + typeExpected.getName(), cause);
-    }
-
-    public PropertyTypeException(String key, Class<?> typeExpected) {
-        this(key, typeExpected, null);
-    }
-
-    public PropertyTypeException(Throwable cause) {
-        super(cause);
-    }
+@Target(ElementType.FIELD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface SensitiveValue {
+    // Has no value
 }